requirements.txt
requests
beepboop
slackclient
slacker
python-dateutil
